Updated parser. Fixed parsing empty spaces after command.
